Fix. For Test Configuration, making TestConfig
