+++ src/test/java/org/prgms/TestConfig.java
@@ -1,0 +1,29 @@
+package org.prgms;
+
+import com.zaxxer.hikari.HikariDataSource;
+import org.springframework.boot.jdbc.DataSourceBuilder;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.jdbc.core.JdbcTemplate;
+
+import javax.sql.DataSource;
+
+@Configuration
+@ComponentScan(basePackages = "org.prgms")
+public class TestConfig {
+    @Bean
+    DataSource dataSource() {
+        return DataSourceBuilder.create()
+                .url("jdbc:mysql://localhost:2215/order_mgmt")
+                .username("test-user")
+                .password("test111!")
+                .type(HikariDataSource.class)
+                .build();
+    }
+
+    @Bean
+    JdbcTemplate jdbcTemplate(DataSource dataSource) {
+        return new JdbcTemplate(dataSource);
+    }
+}
