+++ src/test/java/org/prgms/customer/repository/JdbcCustomerRepositoryTest.java
@@ -1,21 +1,15 @@
 package org.prgms.customer.repository;
 
-import com.zaxxer.hikari.HikariDataSource;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import org.prgms.TestConfig;
 import org.prgms.customer.Customer;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.jdbc.DataSourceBuilder;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.test.context.junit.jupiter.SpringJUnitConfig;
 
-import javax.sql.DataSource;
 import java.util.List;
 import java.util.UUID;
 
@@ -25,27 +19,8 @@ import static com.wix.mysql.config.Charset.UTF8;
 import static com.wix.mysql.config.MysqldConfig.aMysqldConfig;
 import static com.wix.mysql.distribution.Version.v8_0_11;
 
-@SpringJUnitConfig
+@SpringJUnitConfig(value = TestConfig.class)
 class CustomerRepositoryTest {
-
-    @Configuration
-    @ComponentScan(basePackages = "org.prgms")
-    static class Config {
-        @Bean
-        DataSource dataSource() {
-            return DataSourceBuilder.create()
-                    .url("jdbc:mysql://localhost:2215/order_mgmt")
-                    .username("test-user")
-                    .password("test111!")
-                    .type(HikariDataSource.class)
-                    .build();
-        }
-
-        @Bean
-        JdbcTemplate jdbcTemplate(DataSource dataSource) {
-            return new JdbcTemplate(dataSource);
-        }
-    }
 
     @Autowired
     private CustomerRepository jdbcCustomerRepository;
