+++ src/test/java/org/prgms/voucher/service/VoucherServiceTest.java
@@ -5,21 +5,15 @@ import org.hamcrest.Matchers;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import org.prgms.TestConfig;
 import org.prgms.voucher.FixedAmountVoucher;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
 import org.springframework.test.context.junit.jupiter.SpringJUnitConfig;
 
 import java.util.UUID;
 
-@SpringJUnitConfig
+@SpringJUnitConfig(value = TestConfig.class)
 class VoucherServiceTest {
-
-    @Configuration
-    @ComponentScan(basePackages = "org.prgms.voucher")
-    static class Config {
-    }
 
     @Autowired
     private VoucherService voucherService;
